% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first letter of each word}
\usage{
capitalize(x, force_tolower = FALSE)
}
\arguments{
\item{x}{A vector of character strings to be capitalized.}

\item{force_tolower}{Whether to force the remaining letters to be lowercase.}
}
\description{
Capitalize the first letter of each word
}
\examples{
x <- c(
  "hello world",
  "Hello world",
  "hello World"
)
capitalize(x)
}
