% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download}
\alias{download}
\title{Download File from the Internet}
\usage{
download(
  url,
  destfile,
  methods = c("auto", "wget", "libcurl", "curl", "wininet", "internal"),
  max_tries = 2,
  use_httr = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{url}{a \code{\link{character}} string (or longer vector
    for the \code{"libcurl"} method) naming the URL of a resource to be
    downloaded.}

\item{destfile}{a character string (or vector, see the \code{url}
    argument) with the file path where the downloaded file is to be
    saved.  Tilde-expansion is performed.}

\item{methods}{Methods to be used for downloading files.
The default is to try different download methods in turn until the download is successfully completed.}

\item{max_tries}{Number of tries for each download method.}

\item{use_httr}{Logical value, default is \code{FALSE}.
Whether to use \link[httr2:request]{httr2::request} to download the file.}

\item{verbose}{Logical value, default is \code{TRUE}.
Whether to print progress messages.}

\item{...}{Other arguments passed to \link[utils:download.file]{utils::download.file}}
}
\description{
Download File from the Internet
}
