% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.R
\name{figlet_font}
\alias{figlet_font}
\title{Get a figlet font}
\usage{
figlet_font(font)
}
\arguments{
\item{font}{Path or name of the font to load}
}
\value{
A `figlet_font` object for use with [figlet]
}
\description{
Get a figlet font
}
