% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_namespace_fun}
\alias{get_namespace_fun}
\title{Get a function from a namespace}
\usage{
get_namespace_fun(pkg, fun)
}
\arguments{
\item{pkg}{The name of the package.}

\item{fun}{The name of the function.}
}
\value{
Function.
}
\description{
Get a function from a namespace
}
