% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{get_verbose}
\alias{get_verbose}
\title{Get the verbose option}
\usage{
get_verbose(verbose = NULL)
}
\arguments{
\item{verbose}{The verbose option.
Default is `NULL`, which means to get the verbose option from the global options.}
}
\value{
The verbose option.
}
\description{
Get the verbose option from the global options or the local argument.
}
\examples{
get_verbose()
get_verbose(verbose = FALSE)
get_verbose(verbose = TRUE)

options(log_message.verbose = FALSE)
get_verbose()
get_verbose(verbose = TRUE)

options(log_message.verbose = TRUE)
get_verbose()

options(log_message.verbose = NULL)
}
