% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%ss\%}
\alias{\%ss\%}
\title{Value selection operator}
\usage{
a \%ss\% b
}
\arguments{
\item{a}{The left side value to check}

\item{b}{The right side value to use if a is NULL}
}
\value{
a if it is not NULL, otherwise b
}
\description{
This operator returns the left side if it's not NULL,
otherwise it returns the right side.
}
\examples{
NULL \%ss\% 10
5 \%ss\% 10
}
