% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_apple_silicon}
\alias{is_apple_silicon}
\title{Check if the system is running on Apple Silicon}
\usage{
is_apple_silicon()
}
\value{
A logical value.
}
\description{
Check if the system is running on Apple Silicon
}
