% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_depth}
\alias{max_depth}
\title{Maximum depth of a list}
\usage{
max_depth(x, depth = 0)
}
\arguments{
\item{x}{A list.}

\item{depth}{The depth of the list.}
}
\description{
Maximum depth of a list
}
\examples{
x <- list(
  a = list(b = list(c = 1)),
  d = list(e = list(f = 2))
)
max_depth(x)
}
