% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meanp}
\alias{meanp}
\title{Mean p-value}
\usage{
meanp(p)
}
\arguments{
\item{p}{A vector of p-values.}
}
\description{
Mean p-value
}
\examples{
p <- c(0.01, 0.02, 0.03, 0.04, 0.05)
meanp(p)
meanp(p, log.p = TRUE)
}
