% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minimump}
\alias{minimump}
\title{Minimum P-value}
\usage{
minimump(p, alpha = 0.05, log.p = FALSE)
}
\arguments{
\item{p}{A vector of P-values.}

\item{alpha}{The significance level.}

\item{log.p}{Whether to return the log of the P-value.}
}
\description{
Minimum P-value
}
\examples{
p <- c(0.01, 0.02, 0.03, 0.04, 0.05)
minimump(p)
minimump(p, alpha = 0.01)
minimump(p, log.p = TRUE)
}
