% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisutils-package.R
\name{print.logo}
\alias{print.logo}
\title{print logo}
\usage{
\method{print}{logo}(x, ...)
}
\arguments{
\item{x}{Input information.}

\item{...}{Other parameters.}
}
\value{
Print the ASCII logo
}
\description{
print logo
}
