% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_square}
\alias{r_square}
\title{coefficient of determination (\eqn{R^2})}
\usage{
r_square(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector with ground truth values.}

\item{y_pred}{A numeric vector with predicted values.}
}
\value{
\eqn{R^2} value
}
\description{
coefficient of determination (\eqn{R^2})
}
\examples{
r_square(rnorm(100), rnorm(100))
}
