% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r.R
\name{remove_r}
\alias{remove_r}
\title{Check and remove R packages}
\usage{
remove_r(packages, lib = .libPaths()[1], verbose = TRUE)
}
\arguments{
\item{packages}{Package to be removed.}

\item{lib}{The location of the library directories where to remove the packages.}

\item{verbose}{Whether to print the message.
Default is \code{TRUE}.}
}
\description{
Check and remove R packages
}
