% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{rescale}
\alias{rescale}
\title{Rescale numeric vector}
\usage{
rescale(x, from = range(x, na.rm = TRUE, finite = TRUE), to = c(0, 1))
}
\arguments{
\item{x}{A numeric vector.}

\item{from}{The range of the original data.}

\item{to}{The range of the rescaled data.}
}
\value{
A numeric vector with rescaled values.
}
\description{
Rescale numeric vector
}
\examples{
x <- rnorm(10)
rescale(x)
rescale(x, from = c(0, 1))
rescale(x, to = c(0, 2))
}
