% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix_tools.R
\name{row_maxs}
\alias{row_maxs}
\title{Row maxima}
\usage{
row_maxs(x, na.rm = TRUE)
}
\arguments{
\item{x}{A dense or sparse matrix.}

\item{na.rm}{Whether to ignore `NA` values.}
}
\value{
A numeric vector of row maxima. Implicit sparse zeros are
treated as zero.
}
\description{
Row maxima
}
\examples{
row_maxs(matrix(c(-2, 0, 3, 1), nrow = 2))
}
