% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_cor.R
\name{sparse_cov_cor}
\alias{sparse_cov_cor}
\title{Fast correlation and covariance calculation for sparse matrices}
\usage{
sparse_cov_cor(x, y = NULL)
}
\arguments{
\item{x}{Sparse matrix or character vector.}

\item{y}{Sparse matrix or character vector.}
}
\value{
A list with covariance and correlation matrices.
}
\description{
Fast correlation and covariance calculation for sparse matrices
}
\examples{
m1 <- simulate_sparse_matrix(
  100, 100
)
m2 <- simulate_sparse_matrix(
  100, 100,
  sparsity = 0.05
)
a <- sparse_cov_cor(m1, m2)
a$cov[1:5, 1:5]
a$cor[1:5, 1:5]
}
