% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_wrap}
\alias{str_wrap}
\title{Wrap text}
\usage{
str_wrap(x, width = 80)
}
\arguments{
\item{x}{A character vector to wrap.}

\item{width}{The maximum width of the lines.}
}
\value{
A character vector with wrapped text.
}
\description{
Wrap text
}
\examples{
str_wrap(rep("Hello, world!", 10))
str_wrap(rep("Hello, world!", 10), width = 10)
}
