% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkg_file.R
\name{add_pkg_file}
\alias{add_pkg_file}
\title{Add package file}
\usage{
add_pkg_file(
  use_figlet = TRUE,
  figlet_font = "Slant",
  colors = c("red", "yellow", "green", "magenta", "cyan", "yellow", "green", "white",
    "magenta", "cyan"),
  unicode = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{use_figlet}{Whether to use figlet for ASCII art generation.
Default is \code{TRUE}. Details see \link{figlet}.}

\item{figlet_font}{Character string, figlet font to use.
Default is \code{"Slant"}.}

\item{colors}{Character vector, colors to use for the logo elements.}

\item{unicode}{Whether to use Unicode symbols.
Default is \code{TRUE}.}

\item{verbose}{Whether to print progress messages.
Default is \code{TRUE}.}
}
\value{
Creates a file in specified output directory.
}
\description{
Automatically generate a file containing functions and related code for R package development.
}
