% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{check_pkg_status}
\alias{check_pkg_status}
\title{Check if a package is installed with the specified version}
\usage{
check_pkg_status(pkg, version = NULL, lib = .libPaths()[1])
}
\arguments{
\item{pkg}{Package name.}

\item{version}{Package version to check. If \code{NULL}, only checks if the package is installed.}

\item{lib}{The location of the library directories where to install the packages.}
}
\value{
\code{TRUE} if the package is installed with the specified version, \code{FALSE} otherwise.
}
\description{
Check if a package is installed with the specified version
}
