% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{check_r}
\alias{check_r}
\title{Check and install R packages}
\usage{
check_r(
  packages,
  lib = .libPaths()[1],
  dependencies = NA,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{packages}{Package to be installed.
Package source can be \emph{CRAN}, \emph{Bioconductor} or \emph{Github}.
By default, the package name is extracted according to the \code{packages} parameter.}

\item{lib}{The location of the library directories where to install the packages.}

\item{dependencies}{Which dependencies to install.
Passed to \link[pak:pkg_install]{pak::pkg_install}.
Default is \code{NA}, auto install hard dependencies: \emph{Depends}, \emph{Imports},
and \emph{LinkingTo}, excluding \emph{Suggests}.}

\item{force}{Whether to force the installation of packages.
Default is \code{FALSE}.}

\item{verbose}{Whether to print the message.
Default is \code{TRUE}.}
}
\value{
Package installation status.
}
\description{
Check and install R packages
}
