% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r.R
\name{check_r}
\alias{check_r}
\title{Check and install R packages}
\usage{
check_r(
  packages,
  lib = .libPaths()[1],
  dependencies = TRUE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{packages}{Package to be installed.
Package source can be \emph{CRAN}, \emph{Bioconductor} or \emph{Github}.
By default, the package name is extracted according to the \code{packages} parameter.}

\item{lib}{The location of the library directories where to install the packages.}

\item{dependencies}{Whether to install dependencies of the packages.
Default is \code{TRUE}.}

\item{force}{Whether to force the installation of packages.
Default is \code{FALSE}.}

\item{verbose}{Whether to print the message.
Default is \code{TRUE}.}
}
\value{
Package installation status.
}
\description{
Check and install R packages
}
