% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download}
\alias{download}
\title{Download file from the Internet}
\usage{
download(
  url,
  destfile,
  methods = c("auto", "wget", "libcurl", "curl", "wininet", "internal"),
  quiet = FALSE,
  ...,
  max_tries = 2
)
}
\arguments{
\item{url}{a \code{\link{character}} string (or longer vector
    for the \code{"libcurl"} method) naming the URL of a resource to be
    downloaded.}

\item{destfile}{a character string (or vector, see the \code{url}
    argument) with the file path where the downloaded file is to be
    saved.  Tilde-expansion is performed.}

\item{methods}{Methods to be used for downloading files.
Default is \code{"auto"}, which means to try different download methods when the current method fails.}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and
    the progress bar.}

\item{...}{Other arguments passed to \link[utils:download.file]{utils::download.file}.}

\item{max_tries}{Number of tries for each download method.
Default is \code{2}.}
}
\description{
Download file from the Internet
}
