% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figlet.R
\name{list_figlet_fonts}
\alias{list_figlet_fonts}
\title{List available figlet fonts}
\usage{
list_figlet_fonts()
}
\value{
Character vector of available font names.
}
\description{
List all figlet font files available in the package or system.
}
\examples{
list_figlet_fonts()
}
