% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxutils-package.R
\docType{package}
\name{thisutils-package}
\alias{thisutils}
\alias{thisutils-package}
\title{thisutils: inferring cell-type specific gene regulatory network}
\source{
\url{https://github.com/mengxu98/thisutils}
}
\description{
An R package for inferring cell-type specific gene regulatory network from single-cell RNA-seq data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/thisutils/}
  \item Report bugs at \url{https://github.com/mengxu98/thisutils/issues}
}

}
\author{
Meng xu (Maintainer), \email{mengxu98@qq.com}
}
