% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisutils-package.R
\docType{package}
\name{thisutils-package}
\alias{thisutils}
\alias{thisutils-package}
\title{Collection of Utility Functions for Data Analysis and Computing}
\source{
\url{https://mengxu98.github.io/thisutils/}
}
\description{
Provides utility functions for data analysis and scientific computing.
Includes functions for parallel processing, and other computational tasks to streamline workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/thisutils/}
  \item Report bugs at \url{https://github.com/mengxu98/thisutils/issues}
}

}
\author{
Meng Xu (Maintainer), \email{mengxu98@qq.com}
}
