% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisutils-package.R
\docType{package}
\name{thisutils-package}
\alias{thisutils}
\alias{thisutils-package}
\title{thisutils: An R package for utility functions.}
\source{
\url{https://mengxu98.github.io/thisutils/}
}
\description{
An R package for utility functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mengxu98/thisutils}
  \item Report bugs at \url{https://github.com/mengxu98/thisutils/issues}
}

}
\author{
Meng Xu (Maintainer), \email{mengxu98@qq.com}
}
