% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisutils-package.R
\name{thisutils_logo}
\alias{thisutils_logo}
\title{The logo of thisutils}
\usage{
thisutils_logo(unicode = cli::is_utf8_output())
}
\arguments{
\item{unicode}{Unicode symbols on UTF-8 platforms.
Default is \link[cli:is_utf8_output]{cli::is_utf8_output}.}
}
\value{
A character vector with class \code{thisutils_logo}.
}
\description{
The thisutils logo, using ASCII or Unicode characters
Use \link[cli:ansi_strip]{cli::ansi_strip} to get rid of the colors.
}
\examples{
thisutils_logo()
}
\references{
\url{https://github.com/tidyverse/tidyverse/blob/main/R/logo.R}
}
